pyproject.toml
[project]
name = "winix-api"
version = "0.1.0"
description = "Python client for interacting with Winix air purifier devices via the Winix IoT API. Provides authentication, device management, and control."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Chocotonic" }]
dependencies = [
    "black>=25.12.0",
    "pyjwt>=2.10.1",
    "pytest-cov>=7.0.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.0",
    "warrant-lite>=1.0.4",
]
